Bug 1673553 part 43 - Move ObjectGroup ubi code from TypeInference to ObjectGroup files. r=iain
Depends on D98146

Differential Revision: https://phabricator.services.mozilla.com/D98147
